initOperationsAndResultsClass

+++ app/src/main/java/javaclicalculator/Operations.java
@@ -1,8 +1,8 @@
 package javaclicalculator;
 
 public abstract class Operations {
-    abstract double add(int a, int b);
-    abstract double subtract(int a, int b);
-    abstract double multiply(int a, int b);
-    abstract double divide(int a, int b);
+    abstract double add(double a, double b);
+    abstract double subtract(double a, double b);
+    abstract double multiply(double a, double b);
+    abstract double divide(double a, double b);
 }
